feat: implement Study Group feature with complete CRUD operations
- Add Study Group entities (StudyGroup, GroupMember, GroupContent, GroupInvite, GroupMessage)
- Implement repositories, services, and controllers for all Study Group components
- Add security configuration for Study Group endpoints
- Fix migration files order and foreign key references
- Resolve authentication and authorization issues
- Add comprehensive DTOs and exception handling
- Update pom.xml dependencies and application configuration
- Fix linter errors in GroupAuth, GroupMemberServiceImpl, and GroupMessageServiceImpl

+++ src/main/java/se499/kayaanbackend/security/config/JwtAuthenticationFilter.java
@@ -54,7 +54,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
     if (userIdentifier != null && SecurityContextHolder.getContext().getAuthentication() == null) {
       UserDetails userDetails;
       if (userIdentifier.matches("\\d+")) {          // subject is numeric -> treat as id
-        Long id = Long.parseLong(userIdentifier);
+        Integer id = Integer.parseInt(userIdentifier);
         var userOpt = userDao.findById(id);
         if (userOpt.isEmpty()) {
           filterChain.doFilter(request, response);
