+++ src/main/java/se499/kayaanbackend/security/user/UserRepository.java
@@ -1,10 +1,10 @@
 package se499.kayaanbackend.security.user;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Integer> {
 
   Optional<User> findByEmail(String email);
